fix: reconnect on build_failed error
